requirements.txt
# Debt Re-Aging Case Factory
# Python Dependencies

# Core framework
streamlit>=1.28.0

# PDF processing
PyMuPDF>=1.23.0

# Image processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python-headless>=4.8.0  # For advanced image preprocessing

# Templating
Jinja2>=3.1.2

# Data handling
PyYAML>=6.0
pandas>=2.0.0

# Date utilities
python-dateutil>=2.8.2

# PDF generation (optional)
markdown>=3.5.0
# weasyprint>=60.0  # Uncomment for PDF export (requires system dependencies)

# Word document generation
python-docx>=1.1.0

# Fuzzy string matching (for creditor identification)
rapidfuzz>=3.5.0

# API framework (optional, for CMS integration)
flask>=3.0.0
flask-cors>=4.0.0

# HTTP client (for webhooks)
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Monitoring
psutil>=5.9.0

# Note: Tesseract OCR must be installed separately:
# - Windows: https://github.com/UB-Mannheim/tesseract/wiki
# - Mac: brew install tesseract
# - Linux: sudo apt-get install tesseract-ocr

# Note: WeasyPrint (for PDF export) requires system dependencies:
# - Windows: https://doc.courtbouillon.org/weasyprint/stable/first_steps.html#windows
# - Mac: brew install pango
# - Linux: sudo apt-get install libpango-1.0-0 libpangocairo-1.0-0
